Rework adding data set
- Use pointblank for the data dictionary
- Create data set automatically
- Make filename of data set all lowercase

As for the reason: I think it doesn't make
sense to use datawizard for the data
dictionary, as a proper dictionary should
contain information that cannot be just
"calculated" from the data object. For
example, the information about valid values
has to come from the user, rather than from
the data set itself.